core: merge tee_ta_manager{,_unpg}.{c,h}
Merges tee_ta_manager_unpg.h into tee_ta_manager.h and
tee_ta_manager_unpg.c into tee_ta_manager.c

Tested-by: Jens Wiklander <jens.wiklander@linaro.org> (QEMU, FVP)
