+++ core/kernel/sub.mk
@@ -1,6 +1,5 @@
 srcs-y += assert.c
 srcs-y += tee_dispatch.c
 srcs-y += tee_misc.c
-srcs-y += tee_ta_manager_unpg.c
 srcs-y += panic.c
 srcs-y += handle.c
